LayoutTests/animations/CSSKeyframesRule-parameters-expected.txt

Tests that the parameters are mandatory in CSSKeyframesRule API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS keyframes.__proto__ is CSSKeyframesRule.prototype
PASS keyframes.appendRule() threw exception TypeError: Not enough arguments.
PASS keyframes.deleteRule() threw exception TypeError: Not enough arguments.
PASS keyframes.findRule() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
